Refactor: shared types and helpers for report markdown
Address review feedback on the markdown report route:

- Revert API_URL default to /api; derive an absolute base URL on the
  server from incoming request headers via a new getServerApiUrl()
  helper so deployments don't depend on a localhost backend.
- Hoist SimpleReport into src/types/report.ts so the page fetch and
  markdown generator share a single source of truth.
- Extract Issue/groupIssues/describeOccurrence into
  src/utils/issue-grouping.ts so DetailedScoreSection and the markdown
  report use the same grouping/sorting rules.
