+++ apps/web/src/types/report.ts
@@ -1,0 +1,12 @@
+export type SimpleReport = {
+  docsScore: number;
+  completenessScore: number;
+  score: number;
+  securityScore: number;
+  sdkGenerationScore: number;
+  fileExtension: "json" | "yaml";
+  title: string;
+  version: string;
+  shortSummary?: string;
+  longSummary?: string;
+};
