+++ apps/web/src/utils/env.ts
@@ -1,5 +1,1 @@
-const isServer = typeof window === "undefined";
-
-export const API_URL =
-  process.env.NEXT_PUBLIC_API_URL ??
-  (isServer ? "http://localhost:3001" : "/api");
+export const API_URL = process.env.NEXT_PUBLIC_API_URL ?? "/api";
