+++ apps/web/src/utils/generate-markdown-report.ts
@@ -1,26 +1,11 @@
-import type { RatingOutput, SpectralReport } from "@rate-my-openapi/core";
+import type { RatingOutput } from "@rate-my-openapi/core";
 
-type SimpleReport = {
-  docsScore: number;
-  completenessScore: number;
-  score: number;
-  securityScore: number;
-  sdkGenerationScore: number;
-  fileExtension: "json" | "yaml";
-  title: string;
-  version: string;
-  shortSummary?: string;
-  longSummary?: string;
-};
-
-type Issue = SpectralReport[number];
-
-type GroupedIssue = {
-  code: string;
-  severity: number;
-  message: string;
-  occurrences: Issue[];
-};
+import type { SimpleReport } from "@/types/report";
+import {
+  describeOccurrence,
+  groupIssues,
+  type Issue,
+} from "@/utils/issue-grouping";
 
 const SEVERITY_LABEL: Record<number, string> = {
   0: "Error",
@@ -28,59 +13,6 @@ const SEVERITY_LABEL: Record<number, string> = {
   2: "Info",
   3: "Hint",
 };
-
-const HTTP_METHODS = new Set([
-  "get",
-  "post",
-  "put",
-  "patch",
-  "delete",
-  "options",
-  "head",
-  "trace",
-]);
-
-function describeOccurrence(issue: Issue): string {
-  const parts = issue.path.map(String);
-  const pathsIdx = parts.indexOf("paths");
-  if (pathsIdx >= 0 && pathsIdx + 2 < parts.length) {
-    const route = parts[pathsIdx + 1];
-    const method = parts[pathsIdx + 2];
-    if (HTTP_METHODS.has(method.toLowerCase())) {
-      return `${method.toUpperCase()} ${route}`;
-    }
-  }
-  return parts.join(".") || "(root)";
-}
-
-function groupIssues(issues: Issue[]): GroupedIssue[] {
-  const map = new Map<string, GroupedIssue>();
-  for (const issue of issues) {
-    const code = String(issue.code);
-    const existing = map.get(code);
-    if (existing) {
-      existing.occurrences.push(issue);
-      if (issue.severity < existing.severity) {
-        existing.severity = issue.severity;
-        existing.message = issue.message;
-      }
-    } else {
-      map.set(code, {
-        code,
-        severity: issue.severity,
-        message: issue.message,
-        occurrences: [issue],
-      });
-    }
-  }
-  return Array.from(map.values()).sort((a, b) => {
-    if (a.severity !== b.severity) return a.severity - b.severity;
-    if (a.occurrences.length !== b.occurrences.length) {
-      return b.occurrences.length - a.occurrences.length;
-    }
-    return a.message.localeCompare(b.message);
-  });
-}
 
 function renderCategory(
   title: string,
